+++ packages/browser-wallet/src/popup/pages/Account/Earn/Baking/FlowPages/MetadataUrl.tsx
@@ -15,6 +15,7 @@ type MetadataUrlForm = {
 type MetadataUrlProps = MultiStepFormPageProps<ConfigureBakerFlowState['metadataUrl'], ConfigureBakerFlowState>;
 
 export function MetadataUrlPage({ initial, onNext }: MetadataUrlProps) {
+    const { t: tShared } = useTranslation('shared');
     const { t } = useTranslation('account', { keyPrefix: 'baking.configure' });
     const defaultValues: Partial<MetadataUrlForm> = useMemo(
         () => (initial ? { url: initial } : { url: '' }),
@@ -43,7 +44,7 @@ export function MetadataUrlPage({ initial, onNext }: MetadataUrlProps) {
                         />
                     </div>
                     <Submit className="m-t-20" width="wide">
-                        {t('continueButton')}
+                        {tShared('continue')}
                     </Submit>
                 </>
             )}